Establish package identity and quality enforcement
