+++ .github/workflows/quality.yml
@@ -1,0 +1,67 @@
+name: Quality
+
+on:
+  pull_request:
+  push:
+    branches: [main]
+
+permissions:
+  contents: read
+
+concurrency:
+  group: quality-${{ github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  quality:
+    runs-on: ubuntu-latest
+    timeout-minutes: 15
+
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v7
+
+      - name: Set up Python
+        uses: actions/setup-python@v6
+        with:
+          python-version: "3.12"
+
+      - name: Set up uv
+        uses: astral-sh/setup-uv@v9
+        with:
+          version: "0.11.33"
+          enable-cache: true
+
+      - name: Install locked dependencies
+        run: uv sync --locked --all-groups
+
+      - name: Check lockfile
+        run: uv lock --check
+
+      - name: Lint
+        run: uv run ruff check .
+
+      - name: Check formatting
+        run: uv run ruff format --check .
+
+      - name: Type check
+        run: uv run mypy
+
+      - name: Find certain dead code
+        run: uv run vulture
+
+      - name: Report complexity
+        run: |
+          uv run radon cc src/echoflow --total-average
+          uv run radon mi src/echoflow
+
+      - name: Test with branch coverage
+        run: >-
+          uv run pytest
+          --cov=echoflow
+          --cov-branch
+          --cov-report=term-missing
+          --cov-report=xml
+
+      - name: Build distributions
+        run: uv build
