+++ docs/development/testing-and-bisect.md
@@ -6,11 +6,11 @@ Every capability keeps tests in a `tests/` directory directly beneath the
 package it protects:
 
 ```text
-src/core/health_check.py
-src/core/tests/test_health_check.py
+src/echoflow/core/health_check.py
+src/echoflow/core/tests/test_health_check.py
 
-src/interfaces/local_file_manager.py
-src/interfaces/tests/test_local_file_manager.py
+src/echoflow/interfaces/local_file_manager.py
+src/echoflow/interfaces/tests/test_local_file_manager.py
 ```
 
 Repository-wide pytest configuration belongs in root `conftest.py`. Shared
@@ -22,19 +22,20 @@ packages are excluded from built distributions.
 Use the smallest trustworthy oracle first, then widen it:
 
 1. One test node:
-   `uv run pytest src/core/tests/test_health_check.py::test_name`
+   `uv run pytest src/echoflow/core/tests/test_health_check.py::test_name`
 2. One capability suite:
-   `uv run pytest src/core/tests/test_health_check.py`
+   `uv run pytest src/echoflow/core/tests/test_health_check.py`
 3. One package suite:
-   `uv run pytest src/core/tests`
+   `uv run pytest src/echoflow/core/tests`
 4. Tests that failed most recently:
    `uv run pytest --last-failed`
 5. Full behavioral suite:
    `uv run pytest`
 6. Full branch coverage:
-   `uv run pytest --cov=src --cov-branch --cov-report=term-missing`
+   `uv run pytest --cov=echoflow --cov-branch --cov-report=term-missing`
 7. Static and metric gates:
-   `uv run ruff check .`, `uv run vulture`, and `uv run radon cc src --total-average`
+   `uv run ruff check .`, `uv run ruff format --check .`, `uv run mypy`,
+   `uv run vulture`, and `uv run radon cc src/echoflow --total-average`
 8. Targeted mutation scope for changed decision logic, followed by package
    build and lock verification.
 
@@ -46,7 +47,7 @@ let Git invoke the smallest test that proves the regression:
 
 ```bash
 git bisect start BAD GOOD
-git bisect run uv run pytest src/core/tests/test_health_check.py -q
+git bisect run uv run pytest src/echoflow/core/tests/test_health_check.py -q
 git bisect reset
 ```
