--- src/app/__init__.py
@@ -1,4 +1,0 @@
-# src/app/__init__.py
-from src.app.app_container import AppContainer
-
-__all__ = ["AppContainer"]
