--- src/core/ilogger.py
@@ -1,12 +1,0 @@
-# src/core/ilogger.py
-
-from typing import Protocol, runtime_checkable
-
-
-@runtime_checkable
-class ILogger(Protocol):
-    def debug(self, message: str, **kwargs): ...
-    def info(self, message: str, **kwargs): ...
-    def warning(self, message: str, **kwargs): ...
-    def error(self, message: str, **kwargs): ...
-    def bind(self, **kwargs): ...
