+++ src/echoflow/__init__.py
@@ -1,0 +1,1 @@
+"""EchoFlow local audio processing application."""
