+++ src/echoflow/__main__.py
@@ -1,0 +1,6 @@
+"""Module entry point for ``python -m echoflow``."""
+
+from echoflow.cli import main
+
+if __name__ == "__main__":
+    main()
