+++ src/echoflow/app/__init__.py
@@ -1,0 +1,5 @@
+"""Application composition."""
+
+from echoflow.app.app_container import AppContainer
+
+__all__ = ["AppContainer"]
