+++ src/echoflow/core/ilogger.py
@@ -1,0 +1,10 @@
+from typing import Protocol, Self, runtime_checkable
+
+
+@runtime_checkable
+class ILogger(Protocol):
+    def debug(self, message: str, **kwargs: object) -> None: ...
+    def info(self, message: str, **kwargs: object) -> None: ...
+    def warning(self, message: str, **kwargs: object) -> None: ...
+    def error(self, message: str, **kwargs: object) -> None: ...
+    def bind(self, **kwargs: object) -> Self: ...
