+++ src/echoflow/core/performance_tracker.py
