+++ src/echoflow/interfaces/__init__.py
@@ -1,0 +1,4 @@
+from echoflow.interfaces.base_file_manager import FileManager, FileMetadata
+from echoflow.interfaces.local_file_manager import LocalFileManager
+
+__all__ = ["FileManager", "FileMetadata", "LocalFileManager"]
