+++ src/echoflow/interfaces/base_file_manager.py
