--- src/interfaces/__init__.py
@@ -1,4 +1,0 @@
-from src.interfaces.base_file_manager import FileManager, FileMetadata
-from src.interfaces.local_file_manager import LocalFileManager
-
-__all__ = ["FileManager", "FileMetadata", "LocalFileManager"]
