--- src/main.py
@@ -1,6 +1,0 @@
-"""Compatibility entry point for `python -m src.main`."""
-
-from src.cli import main
-
-if __name__ == "__main__":
-    main()
